extract burger common data types to file data-types
